requirements.txt
# local package
-e .

# external requirements
numpy~=1.26.0
scikit-learn~=1.3.1
pandas~=2.1.1
pytest~=7.4.2
catboost~=1.2.2
marshmallow-dataclass==8.5.8
marshmallow==3.17.0
mlflow==2.5.0
dvc>=2.0.0
shap==0.39.0
fastapi==0.92.0
uvicorn==0.20.0
httpx==0.24.1

flake8
python-dotenv>=0.5.1
prometheus-fastapi-instrumentator
prometheus-client
joblib~=1.3.2
dataclasses~=0.8
yaml~=0.2.5
pyyaml~=6.0.1
setuptools~=68.0.0
boto3~=1.29.1